+++ Ranger/Services/SharkScope/Metadata/MetadataRequest.swift
@@ -9,6 +9,12 @@
 import Foundation
 
 
+/// 3.1.1. METADATA
+/// Requests metadata information from the server. If the hash parameter
+/// is specified and the value provided is the latest metadata hash, the
+/// response is set to cache on the client. The response header will not
+/// contain an appVersion attribute even if there is a valid application
+/// version.
 struct MetadataRequest: Request
 {