+++ Ranger/Services/SharkScope/PlayerSummary/PlayerSummary.swift
@@ -21,6 +21,37 @@ struct PlayerSummaryResponse: Response
     let timestamp: StringFor<Date>
     let success: StringFor<Bool>
 
-    let PlayerResponse: PlayerResponse
+    let PlayerResponse: PlayerSummaryResponse.PlayerResponse
     let UserInfo: UserInfo
+    
+    
+    struct PlayerResponse: Decodable
+    {
+
+
+        let PlayerView: PlayerView
+
+
+        struct PlayerView: Decodable
+        {
+
+
+            let Filter: String
+            let Player: Player
+            
+
+            struct Player: Decodable
+            {
+
+
+                let country: String
+                let countryName: String
+                let lastActivity: StringFor<Date>
+                let name: String
+                let network: String
+                let Statistics: Statistics
+                // RecentTournaments
+            }
+        }
+    }
 }