+++ Ranger/Services/SharkScope/PlayerSummary/PlayerSummaryRequest.swift
@@ -9,6 +9,10 @@
 import Foundation
 
 
+/// 3.3.1. SUMMARY
+/// Requests player summary information. The response contains basic
+/// information about the user as well as all free statistical
+/// information and a small number of most recent tournament results.
 struct PlayerSummaryRequest: Request
 {